requirements.txt
torch>=1.10.0
torchvision>=0.11.0
torchaudio>=0.10.0
matplotlib>=3.3.4
seaborn>=0.11.1
scikit-learn>=0.24.2
opencv-python>=4.5.3.56
Pillow>=8.3.1
tqdm>=4.62.3
jupyter>=1.0.0
ipykernel>=6.0.0
# For XAI
pytorch-gradcam>=0.2.1
grad-cam
captum>=0.6.0
